Initial performance increase.

+++ src/main/java/com/lothrazar/cyclic/base/ScreenBase.java
@@ -72,13 +72,13 @@ public abstract class ScreenBase<T extends Container> extends ContainerScreen<T>
 
   /**
    * Translate the block name; and draw it in the top center
-   * 
+   *
    * @param name
    */
   protected void drawName(MatrixStack ms, String name) {
     name = UtilChat.lang("block." + ModCyclic.MODID + "." + name);
     drawString(ms, name,
-        (this.getXSize() - this.font.getStringWidth(name)) / 2,
+        (float) (this.getXSize() - this.font.getStringWidth(name)) / 2,
         6.0F);
   }
